repo components restructured
