+++ src/components/ByLangRepos.js
@@ -1,33 +1,12 @@
-import React, { useState, useEffect } from "react";
-import Axios from "axios";
-import Loader from "./Loader";
+import React from "react";
 import RepoNo from "./RepoNo";
 import RepoLayout from "./RepoLayout";
-import { allRepos } from "../Constants";
 import "../styles/repoStyles.scss";
 
 export default function ByLangRepos(props) {
+  const nofork = props.nofork;
   const selectedValue = props.selectedValue;
-  const [repos, setRepos] = useState([]);
-  const [loading, setLoading] = useState(true);
 
-  useEffect(() => {
-    getAllRepos(); // eslint-disable-next-line
-  }, []);
-
-  const sleep = (milliseconds) => {
-    return new Promise((resolve) => setTimeout(resolve, milliseconds));
-  };
-
-  const getAllRepos = async (milliseconds = 200) => {
-    await sleep(milliseconds);
-    Axios.get(`${allRepos}`).then((response) => {
-      setRepos(response.data);
-      setLoading(false);
-    });
-  };
-
-  const nofork = repos.filter((repo) => repo.fork === false);
   const filtered = nofork.filter(
     (repo) => repo.language === `${selectedValue}`
   );
@@ -41,7 +20,6 @@ export default function ByLangRepos(props) {
         {filtered.map((repo) => (
           <RepoLayout key={repo.id} {...repo} />
         ))}
-        {loading && <Loader />}
       </div>
       <section style={{ minHeight: "300px" }}></section>
     </div>
