+++ src/styles/LoaderStyles.scss
@@ -6,8 +6,8 @@
   max-width: 400px;
   min-width: 150px;
   position: relative;
-  top: 70%;
-  left: 30%;
+  top: 30%;
+  left: 20%;
   transform: translate(-50%, -50%);
   font-size: 60px;
   color: #202428;
